feat: implement timer driver
The timer driver lives in the scheduler since this is the
place where we need it the most. We're still on a cooperative
multitasking model for now, since we cannot really switch in
the timer interrupt handler given that we don't save the
full context, so, for now, the only real purpose of the timer
seems to be interrupting the idle task.