doc: explain how linker script forces identity mapping
Generally speaking, because we cannot relocate, `.rodata` references
will be wrong after we try to switch the VM code.

The same issue will apply to static variables but it's possible to
refactor the kernel to work around this.

In any case, making the kernel relocatable is out of scope for me and
for now, so I am documeting this technical decision.