feat(exec): create kernel shadow in user space
The kernel memory map is immutable after boot since we create
an identity map of the whole RAM and keep using it.

Thus, we can clone the kernel map into the user map by rerunning
the same initialization process.

The mapping is there but the permissions are such that userspace
cannot actually use such a mapping. The kernel can, though.

This means that during a trap initiated from userspace, the
ttbr0 we are using works in kernel space.

Let us remember that the main reason for doing this is that we're
not able to remap the kernel to high memory, hence we cannot use
the ttbr1 for higher addresses for the kernel.