feat(vm, trap): save/restore TTBR0_EL1
We need to ensure that we keep a consistent page table and that
when in kernel we use the kernel page table.

Hence, let us add the page table to the trap frame, and let
us switch to the kernel page table once we saved the original
page table.

This means the kernel always sees its own page table and
each process also only sees its own page table.

But, since we map kernel memory (immutable after boot) into
the userspace page table, with appropriate protection, it
means that any trap occuring when running in user space is
not going to fault.