fix(sched): avoid use-after-stack-free
We were creating a thread and passing it a value allocated on
the stack, and this is very wrong to do.

Instead, just transform the current thread into a process.

Much cleaner and much simpler. As a bonus, this means that the
thread #1 becomes the user process #1, very nice.

(If we had to create kernel threads, this would not be the
case anymore, so in the future we will need to allocate unique
IDs also to processes or maybe create a suspended thread and
only start it when the time is ripe, IDK.)

While there, remove code for testing threads and just focus
on booting to userspace here.