refactor: introduce __duration64_t type
This type explicitly refers to a duration expressed using 64 bits.

We cannot use 32 bits for durations anyway.

This diff is part of a series of diffs where the overall objective
is to ban uint64_t and int64_t from code that is portable and instead
use semantically clear equally sized integer types.