fix(sched): sched_proc_start -> sched_proc_exec
The sched_proc_start function was executing the process without
returnning. So, make this explicit by changing the name and by
clearly marking the function as `[[noreturn]]`.

Incidentally, this removes one stray `int64_t` from portable
code, which is what I am trying to do to reduce confusion and
ensure we always know which types we're actually using.